Bugfix: Call "ready" callback when web view is actually ready
There were problems loading multiple views in the same UIViewController.
The first time it loaded in the app, only 1 of the views would load; the others would remain blank.

I'm guessing the JS wasn't fully loaded when `webViewDidLoad:` was called, and so `setContentHTML` wasn't doing anything.

This way, we call the "ready" callback with JS, so it should be after everything is loaded.